Use long for Modifier Flags Internally
comp: Modifier flags are now internally represented as `long`s.

+++ src/compiler/dyvilx/tools/compiler/ast/attribute/Attribute.java
@@ -20,7 +20,7 @@ public interface Attribute extends ASTNode, Resolvable
 {
 	@Nullable IType getType();
 
-	int flags();
+	long flags();
 
 	@Override
 	void resolveTypes(MarkerList markers, IContext context);
